refactor(cli): split oversized publicApi files under line limits
