+++ src/core/publicApi/flagSpecs.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 
 import { flagKind, type FlagKind, type JsonSchema } from "./flagKind.js";
+import { flagNameOf, flagUsage, optionKeyOf } from "./flagNames.js";
 import { type ObjectShape, toObjectShape } from "./objectShape.js";
 
 export type FlagSpec = {
@@ -20,35 +21,6 @@ export type FlagSpec = {
 export type FlagSpecsResult =
   | { ok: true; flags: FlagSpec[] }
   | { ok: false; field: string; reason: string };
-
-const kebabCase = (segment: string): string =>
-  segment.replace(/[A-Z]/g, (char) => `-${char.toLowerCase()}`);
-
-const capitalize = (segment: string): string =>
-  `${segment.slice(0, 1).toUpperCase()}${segment.slice(1)}`;
-
-const optionKeyOf = (path: string[]): string =>
-  path
-    .map((segment, index) => (index === 0 ? segment : capitalize(segment)))
-    .join("");
-
-// The bare flag name, e.g. "--environment-id". `flagUsage` appends a
-// kind-specific value placeholder ("", " <value>", " <values...>", or
-// " <KEY=VALUE...>") on top of this, so two fields of different kinds that
-// share a name produce different `flag` strings even though Commander
-// registers them under the same option. Collision detection keys on this
-// name, not on `flag`, so it still catches that case.
-function flagNameOf(path: string[]): string {
-  return `--${path.map(kebabCase).join("-")}`;
-}
-
-function flagUsage(path: string[], kind: FlagKind): string {
-  const name = flagNameOf(path);
-  if (kind === "boolean") return name;
-  if (kind === "string-array") return `${name} <values...>`;
-  if (kind === "key-value-record") return `${name} <KEY=VALUE...>`;
-  return `${name} <value>`;
-}
 
 // An enum field accepts nothing but its own values, so --help lists them rather
 // than leaving a caller to learn them from a rejected invocation. The wording
