+++ src/domains/publicApi/handle.test.ts
@@ -11,10 +11,8 @@ import {
 } from "~/shell/platform/createPlatformClient.testUtils.js";
 import type { UI } from "~/shell/ui/index.js";
 
-import { buildFlagSpecs } from "~/core/publicApi/flagSpecs.js";
-
 import { buildCommandSpecs, type CommandSpec } from "./commandSpecs.js";
-import { assembleInput, handlePublicApiCommand } from "./handle.js";
+import { handlePublicApiCommand } from "./handle.js";
 
 afterEach(() => {
   mock.restore();
@@ -53,42 +51,6 @@ function ctxWith(ui: UI, platformClient: PlatformClient): AuthCommandContext {
     platformClient,
   };
 }
-
-describe("assembleInput", () => {
-  it("rebuilds nested input from flattened option keys", () => {
-    const contract = {
-      name: "fake.reportStatus",
-      kind: "write",
-      description: "Synthetic nested contract.",
-      input: z.object({
-        environment: z.union([
-          z.strictObject({ id: z.string() }),
-          z.strictObject({ name: z.string() }),
-        ]),
-        externalId: z.string(),
-        metadata: z.object({ commitSha: z.string() }).optional(),
-      }),
-      output: z.object({}),
-    };
-    const flags = buildFlagSpecs(contract.input);
-    if (!flags.ok) throw new Error(flags.reason);
-
-    const assembled = assembleInput(flags.flags, {
-      environmentName: "preview-42",
-      externalId: "vercel_dpl_123",
-      metadataCommitSha: "abc123",
-    });
-
-    expect(assembled).toEqual({
-      ok: true,
-      input: {
-        environment: { name: "preview-42" },
-        externalId: "vercel_dpl_123",
-        metadata: { commitSha: "abc123" },
-      },
-    });
-  });
-});
 
 describe("handlePublicApiCommand", () => {
   it("assembles the input, calls the contract, and outputs the result", async () => {
